Create LICENSE.md
The Technology and Education: A Multidisciplinary Journal (TEMJ) is a monthly open-access, peer-reviewed publication based in the Philippines. It is dedicated to achieving excellence across various facets of formal education. With a commitment to international outreach, TEMJ disseminates regular papers and special issues covering topics of significant interest to educational researchers worldwide.

Employing double-blind reviews, TEMJ ensures impartial evaluation, where reviewers remain unaware of the author(s) of each manuscript. This rigorous process underscores the journal's dedication to scholarly integrity and the advancement of educational discourse on a global scale.